+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # v2.14.42
 
 * added `terminating` property to `TerminableThread`
+* added `Closeable.closed`